+++ src/main/java/com/github/yury1991/PayrollWeb/models/Profit.java
@@ -1,5 +1,7 @@
 package com.github.yury1991.PayrollWeb.models;
 
+import java.math.BigDecimal;
+
 public interface Profit {
-	public double getProfit();	
+	public BigDecimal getProfit();	
 }
